tests/wpt/tests/webxr/depth-sensing/depth_sensing_attributes_throw_not_enabled.https.html

<!DOCTYPE html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="../resources/webxr_util.js"></script>
<script src="../resources/webxr_test_constants.js"></script>

<script>

const testFunctionDepthUsage = function (session, controller, t) {
    t.step(() => {
      assert_throws_dom("InvalidStateError", () => session.depthUsage,
                        "session.depthUsage should throw when depth sensing is disabled");
    });

  return Promise.resolve();
};

const testFunctionDepthDataFormat = function (session, controller, t) {
    t.step(() => {
      assert_throws_dom("InvalidStateError", () => session.depthDataFormat,
                        "session.depthDataFormat should throw when depth sensing is disabled");
    });

  return Promise.resolve();
};

const testFunctionDepthType = function (session, controller, t) {
    t.step(() => {
      assert_throws_dom("InvalidStateError", () => session.depthType,
                        "session.depthType should throw when depth sensing is disabled");
    });

  return Promise.resolve();
};

xr_session_promise_test(
  "XRSession.depthUsage rejects if depth sensing is not enabled on a session",
  testFunctionDepthUsage,
  IMMERSIVE_AR_DEVICE,
  'immersive-ar');

xr_session_promise_test(
  "XRSession.depthDataFormat rejects if depth sensing is not enabled on a session",
  testFunctionDepthDataFormat,
  IMMERSIVE_AR_DEVICE,
  'immersive-ar');

xr_session_promise_test(
  "XRSession.depthType rejects if depth sensing is not enabled on a session",
  testFunctionDepthType,
  IMMERSIVE_AR_DEVICE,
  'immersive-ar');

</script>
