third_party/blink/web_tests/platform/mac-mac15-arm64/external/wpt/webxr/depth-sensing/depth_sensing_attributes_throw_not_enabled.https-expected.txt

This is a testharness.js-based test.
[FAIL] XRSession.depthUsage rejects if depth sensing is not enabled on a session - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession.depthUsage rejects if depth sensing is not enabled on a session - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession.depthDataFormat rejects if depth sensing is not enabled on a session - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession.depthDataFormat rejects if depth sensing is not enabled on a session - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession.depthType rejects if depth sensing is not enabled on a session - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession.depthType rejects if depth sensing is not enabled on a session - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
